requirements.txt
asgiref==3.8.1
Django==4.2.11
django-debug-toolbar==4.3.0
djangorestframework==3.14.0
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
pytest-django==4.9.0
pytz==2024.2
sqlparse==0.5.1
